CORE-2841 Added implicit boolean to DataObject.
Added a boolean field to DataObject for the new is_implicit column in the dataobjects table.
Added OutputPropertyFormatter, and a Unit Test class, for the PropertyFormatterFactory (used by FAPI jobs).
Updated CondorJobRequestFormatter to check for and format implicit Output Properties.
These formatters will return null if the Output DataObject is implicit, so these arguments will not be added to the command line.