CORE-3012 Updated copyAnalysis to also import App.
Refactored the copyAnalysis method in AnalysisEditService so that it will also save the copied App into the database via the WorkflowImportService, in addition to saving it in the OSM via Zoidberg.
Added a setWorkflowImportService method for controllers or calling service layers (aka Donkey).
Refactored some common copy-app code from the editAnalysis method (when the App does not already exist in the user's workspace) into an overloaded copyAnalysis method used by both the edit and copy service calls.