CORE-3115 Added deployed component search service.
Added a searchComponents method to DeployedComponentRetrievalService that will search the deployed_components table by name or description for the given search term.
Added a WorkflowElementSearchService class with a searchDeployedComponents method that uses the new DeployedComponentRetrievalService::searchComponents method.
Refactored BaseWorkflowElementRetrievalService to allow DeployedComponentRetrievalService protected access to some fields and the marshall method.