CORE-4626 Fix for deployed component search endpoint.
Removed the 'escape' clause from the 'where' search clauses, since it was causing an "org.postgresql.util.PSQLException: The column index is out of range: 2, number of columns: 1." exception.
The backslash was probably not escaped properly, but backslash is already the default escape character, so it's not required.
Added support for '*' and '?' as client wildcard search characters.